[clean]

+++ takao/field_type.cpp
@@ -296,7 +296,7 @@ std::string field_type::get_answer() const
         prev_nth = current_nth;
         process_count++;
     }
-    for(int i = prev_nth;i < provided_stones; i++)result.append("\r\n");
+    for(size_t i = prev_nth;i < provided_stones; i++)result.append("\r\n");
     return result;
 }
